+++ llm_gent/core/tools/builtin/__init__.py
@@ -5,6 +5,8 @@ from .file import FileReadTool, FileWriteTool
 from .http import HTTPFetchTool
 from .learn import RecallTool, RememberTool
 from .shell import ShellTool
+from .web_fetch import WebFetchTool
+from .web_search import WebSearchTool
 
 
 __all__ = [
@@ -15,4 +17,6 @@ __all__ = [
     "RecallTool",
     "RememberTool",
     "ShellTool",
+    "WebFetchTool",
+    "WebSearchTool",
 ]